requirements.txt
fastapi
uvicorn[standard]
faiss-cpu
sentence-transformers
numpy
scikit-learn
psycopg2-binary
tqdm_loggable
pika
debugpy
